refactor: consolidate KeygenType imports and remove legacy references
- Introduced a new KeygenType enum in the core/mpc/keygen/KeygenType.ts file.
- Updated multiple files to import KeygenType from the new location, enhancing code clarity.
- Removed obsolete tssType file that previously defined KeygenType, streamlining the codebase.
