+++ clients/desktop/src/vault/keygen/join/JoinKeygenPage.tsx
@@ -1,3 +1,4 @@
+import { KeygenType } from '@core/mpc/keygen/KeygenType'
 import { fromLibType } from '@core/mpc/types/utils/libType'
 import { match } from '@lib/utils/match'
 import { useTranslation } from 'react-i18next'
@@ -14,7 +15,6 @@ import { MpcLibProvider } from '../../../mpc/state/mpcLib'
 import { useAppPathState } from '../../../navigation/hooks/useAppPathState'
 import { useNavigateBack } from '../../../navigation/hooks/useNavigationBack'
 import { CurrentHexEncryptionKeyProvider } from '../../setup/state/currentHexEncryptionKey'
-import { KeygenType } from '../KeygenType'
 import { JoinKeygenSessionStep } from '../shared/JoinKeygenSessionStep'
 import { CurrentServiceNameProvider } from '../shared/state/currentServiceName'
 import { CurrentKeygenTypeProvider } from '../state/currentKeygenType'
