+++ clients/desktop/src/vault/keygen/shared/KeygenProgressIndicator.tsx
@@ -1,3 +1,4 @@
+import { KeygenType } from '@core/mpc/keygen/KeygenType'
 import { match } from '@lib/utils/match'
 import { useTranslation } from 'react-i18next'
 
@@ -5,7 +6,6 @@ import RingProgress from '../../../components/ringProgress/RingProgress'
 import { VStack } from '../../../lib/ui/layout/Stack'
 import { ValueProp } from '../../../lib/ui/props'
 import { Text } from '../../../lib/ui/text'
-import { KeygenType } from '../KeygenType'
 import { useCurrentKeygenType } from '../state/currentKeygenType'
 import { KeygenStatus } from './MatchKeygenSessionStatus'
