+++ clients/desktop/src/vault/keygen/state/currentKeygenType.ts
@@ -1,5 +1,6 @@
+import { KeygenType } from '@core/mpc/keygen/KeygenType'
+
 import { getValueProviderSetup } from '../../../lib/ui/state/getValueProviderSetup'
-import { KeygenType } from '../KeygenType'
 
 export const {
   useValue: useCurrentKeygenType,
