+++ clients/desktop/src/vault/setup/fast/SetupFastVaultPage.tsx
@@ -1,3 +1,5 @@
+import { KeygenType } from '@core/mpc/keygen/KeygenType'
+
 import { Match } from '../../../lib/ui/base/Match'
 import { useStepNavigation } from '../../../lib/ui/hooks/useStepNavigation'
 import { GeneratedMpcLocalPartyIdProvider } from '../../../mpc/localPartyId/state/mpcLocalPartyId'
@@ -9,7 +11,6 @@ import { useDefaultMpcLib } from '../../../mpc/state/defaultMpcLib'
 import { IsInitiatingDeviceProvider } from '../../../mpc/state/isInitiatingDevice'
 import { MpcLibProvider } from '../../../mpc/state/mpcLib'
 import { useNavigateBack } from '../../../navigation/hooks/useNavigationBack'
-import { KeygenType } from '../../keygen/KeygenType'
 import { KeygenFlow } from '../../keygen/shared/KeygenFlow'
 import { GeneratedServiceNameProvider } from '../../keygen/shared/state/currentServiceName'
 import { CurrentKeygenTypeProvider } from '../../keygen/state/currentKeygenType'
