+++ clients/desktop/src/vault/setup/peers/queries/useJoinKeygenUrlQuery.ts
@@ -1,6 +1,7 @@
 import { create, toBinary } from '@bufbuild/protobuf'
 import { toCompressedString } from '@core/chain/utils/protobuf/toCompressedString'
 import { deepLinkBaseUrl } from '@core/config'
+import { KeygenType } from '@core/mpc/keygen/KeygenType'
 import { toLibType } from '@core/mpc/types/utils/libType'
 import { KeygenMessageSchema } from '@core/mpc/types/vultisig/keygen/v1/keygen_message_pb'
 import { ReshareMessageSchema } from '@core/mpc/types/vultisig/keygen/v1/reshare_message_pb'
@@ -13,7 +14,6 @@ import { useTransformQueryData } from '../../../../lib/ui/query/hooks/useTransfo
 import { useMpcServerType } from '../../../../mpc/serverType/state/mpcServerType'
 import { useMpcSessionId } from '../../../../mpc/session/state/mpcSession'
 import { useMpcLib } from '../../../../mpc/state/mpcLib'
-import { KeygenType } from '../../../keygen/KeygenType'
 import { useCurrentServiceName } from '../../../keygen/shared/state/currentServiceName'
 import { useCurrentKeygenType } from '../../../keygen/state/currentKeygenType'
 import { useCurrentVault } from '../../../state/currentVault'
