+++ core/mpc/mpc.ts
@@ -1,9 +1,9 @@
 import { EventsEmit } from '@wailsapp/runtime'
 
 import { DKLS } from './dkls/dkls'
+import { KeygenType } from './keygen/KeygenType'
 import { setKeygenComplete, waitForKeygenComplete } from './keygenComplete'
 import { Schnorr } from './schnorr/schnorrKeygen'
-import { KeygenType } from './tssType'
 /**
  * MPC class - this class handle keygen/reshare using DKLS/Schnorr library
  * NOTE: this class doesn't handle GG20 keygen/keysign/reshare, GG20 is handled by the method defined in golang
