--- core/mpc/tssType.ts
@@ -1,5 +1,0 @@
-export enum KeygenType {
-  Keygen = 'Keygen',
-  Reshare = 'Reshare',
-  Migrate = 'Migrate',
-}
